OAI-61: Deprecate previous converters, refactor abstract converters
