+++ claim_ai/evaluation/converters/legacy_converter/healthcare.py
@@ -1,5 +1,5 @@
 from claim_ai.evaluation import input_models
-from claim_ai.evaluation.converters.legacy_converter.base_converter import AbstractConverter
+from claim_ai.evaluation.converters.base_converter import AbstractConverter
 
 
 class HealthcareServiceConverter(AbstractConverter):
