+++ claim_ai/evaluation/converters/legacy_converter/medical_provisions.py
@@ -1,7 +1,7 @@
 import core
 
 from claim_ai.evaluation import input_models
-from claim_ai.evaluation.converters.legacy_converter.base_converter import AbstractConverter
+from claim_ai.evaluation.converters.base_converter import AbstractConverter
 from claim_ai.evaluation.input_models import ProvidedItem, Claim
