+++ claim_ai/tests/__init__.py
@@ -1,1 +1,1 @@
-from .test_ai_input_converter import *
+from claim_ai.tests.legacy.test_ai_input_converter import *
